chore: update shelljs to 0.8.5 (from 0.8.4)
This fixes GHSA-4rq4-32rv-6wp6 (CVE-2022-0144, 7.1/10 (high)) and  GHSA-64g7-mvw6-v9qj, which was patched in 0.8.5. The vulnerability allowed users of the same system to see output generated by the synchrous version of shell.exec(). This package doesn't use either versions of shell.exec(), but I'm stilling patchting this to avoid the error message generated when you install this package.